Replace with the new implementation
